Added lightning css for minification of css
