+++ src/package.json
@@ -32,7 +32,6 @@
   "dependencies": {
     "async": "^3.2.5",
     "axios": "^1.7.2",
-    "clean-css": "^5.3.3",
     "cookie-parser": "^1.4.6",
     "cross-spawn": "^7.0.3",
     "ejs": "^3.1.10",
@@ -51,9 +50,11 @@
     "jsonminify": "0.4.2",
     "jsonwebtoken": "^9.0.2",
     "languages4translatewiki": "0.1.3",
+    "lightningcss": "^1.25.1",
     "live-plugin-manager": "^1.0.0",
     "lodash.clonedeep": "4.5.0",
     "log4js": "^6.9.1",
+    "lru-cache": "^10.2.0",
     "measured-core": "^2.0.0",
     "mime-types": "^2.1.35",
     "oidc-provider": "^8.4.5",
@@ -75,8 +76,7 @@
     "ueberdb2": "^4.2.63",
     "underscore": "1.13.6",
     "unorm": "1.6.0",
-    "wtfnode": "^0.9.2",
-    "lru-cache": "^10.2.0"
+    "wtfnode": "^0.9.2"
   },
   "bin": {
     "etherpad-healthcheck": "../bin/etherpad-healthcheck",
